pyproject.toml
[project]
name = "database-mcp-server"
version = "1.0.2"
description = "Database MCP Server for SQL operations"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbutils>=3.1.1",
    "fastmcp>=2.11.0",
    "mcp[cli]>=1.12.2",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.1",
]

[project.scripts]
database-mcp-server = "src:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
